Scan-friendly shared-weight memory graph: 14K → 24K tok/s
Replace per-neuron state/msg MLPs with shared-weight cores conditioned
on per-neuron identity embeddings. The key insight: decompose the MLP
input into time-invariant parts (received, primitives, neuron_id, decay)
computed once, and time-varying inject computed for all T via F.linear.
Then fused_scan handles temporal integration.

Architecture changes:
- State MLP: per-neuron [N,H,2D] → shared [H, 3D+1] with base/inject split
- Msg MLP: per-neuron [N,H,2D] → shared [H, 3D] with identity conditioning
- Modulator: stays per-neuron (runs once/segment, negligible cost), H=112
- Triton fused forward kernel for shared-weight pass (no [BS,N,T,D] intermediate)
- Backward via PyTorch reference recomputation (no custom Triton backward)
- Gradient checkpoint on pass 1 to bound VRAM

Preserved: N=512, D=256, K=32 sparse graph, 2-pass frozen messages,
structural plasticity, Hebbian traces, per-token inject/readout,
neuron_id embedding, all gradient paths intact.

+++ src/v8/model.py
@@ -96,7 +96,7 @@ class V8Model(nn.Module):
         # Detach H_mid for memory graph input — lower scan gets its own
         # gradient path through the CE loss, memory graph provides a second
         # gradient channel through inject_memory
-        cc_all = H_mid.detach()  # [BS, T, D]
+        cc_all = H_mid.detach().to(self.memory.dtype)  # [BS, T, D]
         n_segments = T // action_every
 
         mem_out_segs = []
